fixed asset manager added reset

+++ core/src/com/mygdx/gameseven/Constants.java
@@ -5,5 +5,5 @@ import com.badlogic.gdx.Gdx;
 public class Constants {
     public static final float VIEWPORT_WIDTH = 100;
     public static final float VIEWPORT_HEIGHT = 100;
-    public static final float ASPECT_RATIO = Gdx.graphics.getHeight()/Gdx.graphics.getWidth();
+    public static final float ASPECT_RATIO = (float)Gdx.graphics.getHeight()/(float)Gdx.graphics.getWidth();
 }
